// src/app/shared/models/order-dto.ts
import { OrderDetailDto } from "./order-detail-dto";

export class OrderDto {
    securityUsersId : number;
    companyId: number;
    customerId: number;
    orderId: number;
    formNumber: number;
    registerType: string;
    orderTypeId: number;
    orderSourceId: number;
    plantId: number;
    purchaseOrderNumber: string;
    saleTermsId: number;
    placementDate: string | null;
    orderDate: string;
    promisedDate: string | null;
    sampleDeliveryDate: string | null;
    sampleModifiedDate: string | null;
    approvedDate: string | null;
    sampleApprovalOrder: number | null;
    sampleApprovalItem: number | null;
    lastInvoiceDate: string | null;
    lastInvoiceNumber: number | null;
    shipAreaId: number;
    shipViaId: number;
    tradeTermsId: number | null;
    destinationPort: number;
    paymentMethodId: number | null;
    priceListTypeId: number | null;
    salesTaxExceptionId: number | null;
    withholdingTypeId: number | null;
    sampleApprovalTypeId: number | null;
    billingAddress: number | null;
    soldAddress: number | null;
    shippingAddress: number;
    supplierId: number | null;
    supplierPurchaseOrder: number | null;
    intercompanyBillingService: string;
    carrierId: number;
    cusotmerCarrierAccount: string;
    deliveryMethodId: number | null;
    finotexBillingOptionId: number | null;
    billingOptionId: number | null;
    packageTypeId: number | null;
    shippingServiceTypeId: number | null;
    freightCharges: number | null;
    orderStatusId: number | null;
    fulfilledManual: string;
    fulfilledBy: string;
    recordUser: string;
    modifiedByUser: string;
    modifiedDate: string;
    sampleProcessTypeId: number | null;
    sampleUser: string;
    samplePendingToPrint: string;
    sampleLastShipmentDate: string | null;
    sampleLastReceptioDate: string | null;
    sampleShipmentsNumber: number | null;
    suggestedSamplePrice: number | null;
    sampleDeliveryAddress1: string;
    sampleDeliveryAddress2: string;
    samplesNotes: string;
    orderNotes: string;
    detail: Array<OrderDetailDto>;
}