// src/app/shared/models/order-detail-dto.ts
export class OrderDetailDto {
    companyId: number;
    orderId: number;
    orderitem: number;
    requestDate: string;
    promisedDate: string | null;
    promisedDeliveryDate: string | null;
    storeHouseId: number | null;
    productId: string;
    cutId: number;
    adhesiveId: number;
    finishId: number;
    unitMeasureId: string;
    salesUnit: string;
    conversionFactor: number;
    inverseFactor: number;
    priceListId: number | null;
    basicUnitPrice: number;
    markupPercentage: number;
    additionalChargeValue: number;
    chargeByMinimunOrder: number;
    cutPriceByUnit: number;
    unitPrice: number;
    unitPriceWithDiscount: number | null;
    unitCost: number | null;
    quotedUnitPrice: number | null;
    quotedSalePrice: number | null;
    specialPriceFlag: string;
    salePrice: number;
    proformaOrPurchaseUnit: string;
    proformaOrPurchasePrice: number | null;
    artValue: number;
    sampleQuantity: number | null;
    quantityOrdered: number;
    quantityReserved: number;
    quantityPackingList: number;
    quantityInvoiced: number;
    lastInvoiceDate: string | null;
    requestedProduction: number | null;
    productionFinished: number | null;
    productionInProcess: number | null;
    isAuthorizedDelivery: string;
    fulfilledManual: string;
    orderStatusId: number | null;
    modifiedByUser: string;
    modifiedDate: string;
    customerExternalOrder: string;
    customerExternaLote: string;
    customerOrderId: number | null;
    customerItemOrder: number | null;
    loanCodeId: number | null;
    contract: string;
    detailNotes: string;
}